If the user does not have access to an android phone, the app can be run on the emulator which can be accessed through Android Studio. First go to Tools -> Android -> AVD Manager-> Create Virtual Device. From there the user can choose a device that he or she wants to run the app on. Once a device has been created, click the “Run App” button to select a device. Click the button next to ”Launch Emulator” from where the user can select the device that he or she wants to run the app on. Once a device has been chosen, click OK and an emulator will be launched where the app can be run on.
If the user does not have access to an android phone, the app can be run on the emulator which can be accessed through Android Studio. First go to Tools -> Android -> AVD Manager-> Create Virtual Device. From there the user can choose a device that he or she wants to run the app on. Once a device has been created, click the “Run App” button to select a device. Click the button next to ”Launch Emulator” from where the user can select the device that he or she wants to run the app on. Once a device has been chosen, click OK and an emulator will be launched where the app can be run on.